+++ db/migrate/20150922083926_add_server_last_test_finished.rb
@@ -1,0 +1,5 @@
+class AddServerLastTestFinished < ActiveRecord::Migration
+  def change
+    add_column :servers, :last_test_run_date, :datetime
+  end
+end