+++ Changelog.md
@@ -1,7 +1,8 @@
 # Change log
 
-## Unreleased (master)
+## master (unreleased)
 * Fix collapsing navbar on mobile devices (small screens)
+* Add ability to get last test run date for each server
 
 ## 1.4.2
 * Remove executed containers, that no need. They took a lot of space